requirements.txt
transformers>=4.37.2
accelerate>=0.25.0
deepspeed
diffusers>=0.30.1
Pillow
einops
safetensors
timm
tomesd
torchdiffeq
torchsde
xformers
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
SwissArmyTransformer==0.4.12
matplotlib
pytorch_lightning
kornia==0.7.3
wandb==0.17.5
beartype>=0.19.0